backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
httpx>=0.26.0
openai>=1.10.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
